Move most of test_list.py to new cudf classic test directories
