+++ python/cudf/cudf/tests/dataframe/test_constructors.py
@@ -1,0 +1,46 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.testing import assert_eq
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        {"a": [[]]},
+        {"a": [[None]]},
+        {"a": [[1, 2, 3]]},
+        {"a": [[1, 2, 3]], "b": [[2, 3, 4]]},
+        {"a": [[1, 2, 3, None], [None]], "b": [[2, 3, 4], [5]], "c": None},
+        {"a": [[1]], "b": [[1, 2, 3]]},
+        pd.DataFrame({"a": [[1, 2, 3]]}),
+    ],
+)
+def test_df_list_dtypes(data):
+    expect = pd.DataFrame(data)
+    got = cudf.DataFrame(data)
+    assert_eq(expect, got)
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        {"a": [[]]},
+        {"a": [[1, 2, None, 4]]},
+        {"a": [["cat", None, "dog"]]},
+        {
+            "a": [[1, 2, 3, None], [4, None, 5]],
+            "b": [None, ["fish", "bird"]],
+            "c": [[], []],
+        },
+        {"a": [[1, 2, 3, None], [4, None, 5], None, [6, 7]]},
+    ],
+)
+def test_serialize_list_columns(data):
+    df = cudf.DataFrame(data)
+    recreated = df.__class__.deserialize(*df.serialize())
+    assert_eq(recreated, df)
